Done modifications based on comments.
