+++ src/Week1_DataTypes/N1_Fahrenheit.java
@@ -1,10 +1,15 @@
-package Day1_DataTypes;
+package Week1_DataTypes;
 
 import java.util.Scanner;
 
 /**
  * Created by Agnesa on Sep, 2020
+ *
+ * 1. Write a program to convert temperature from Fahrenheit to Celsius degree.
+ * Test Data: Input a degree in Fahrenheit: 212
+ * Expected Output: 212.0 degree Fahrenheit is equal to 100.0 in Celsius
  */
+
 public class N1_Fahrenheit {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
