+++ src/Week1_DataTypes/N2_Inches.java
@@ -1,10 +1,16 @@
-package Day1_DataTypes;
+package Week1_DataTypes;
 
 import java.util.Scanner;
 
 /**
  * Created by Agnesa on Sep, 2020
+ *
+ * 2. Write a program that reads a number in inches, converts it to meters.
+ * Note: One inch is 0.0254 meter.
+ * Test Data: Input a value for inch: 1000
+ * Expected Output: 1000.0 inch is 25.4 meters
  */
+
 public class N2_Inches {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
