+++ src/Week1_DataTypes/N3_Gravity.java
@@ -1,10 +1,14 @@
-package Day1_DataTypes;
+package Week1_DataTypes;
 
 import java.util.Scanner;
 
 /**
  * Created by Agnesa on Sep, 2020
+ *
+ * 3. The force of gravity on the Moon is equal to ~ 17%.
+ * Write a program that by entering your weight will count your weight on the Moon.
  */
+
 public class N3_Gravity {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
