+++ src/Week1_DataTypes/N6_Distance.java
@@ -1,10 +1,15 @@
-package Day1_DataTypes;
+package Week1_DataTypes;
 
 import java.util.Scanner;
 
 /**
  * Created by Agnesa on Sep, 2020
+ * <p>
+ * 6. Write a program that will suggest you to enter the coordinates of 2 cities and will calculate:
+ * a) The distance between 2 cities
+ * b) The middle point between 2 cities
  */
+
 public class N6_Distance {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
@@ -20,7 +25,7 @@ public class N6_Distance {
         double y2 = input.nextDouble();
 
         System.out.println("The Distance Is: ");
-        System.out.println(Math.sqrt((y2 - y1) * (y2 - y1) + (x2 - x1) * (x2 - x1)));
+        System.out.println(Math.pow((y2 - y1), 2));
         System.out.println("Middle Latitude is: " + (x1 + x2) / 2);
         System.out.println("Middle Longitude is: " + (y1 + y2) / 2);
     }
