+++ src/Week1_DataTypes/N8_FloatingPointNumber.java
@@ -1,10 +1,15 @@
-package Day1_DataTypes;
+package Week1_DataTypes;
 
 import java.util.Scanner;
 
 /**
  * Created by Agnesa on Sep, 2020
+ *
+ * 8. Write a program that reads a floating-point number and prints "zero" if the number is zero. Otherwise, print "positive" or "negative". Add "small" if the absolute value of the number is less than 1, or "large" if it exceeds 1,000,000.
+ * Test Data: Input a number: 20.6
+ * Expected Output: positive
  */
+
 public class N8_FloatingPointNumber {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
