--- src/datatypes/DataTypesDemo.java
@@ -1,7 +1,0 @@
-package datatypes;
-
-/**
- * Created by Agnesa on Sep, 2020
- */
-public class DataTypesDemo {
-}
